+++ src/es/taikutsu/src/eu/kanade/tachiyomi/extension/es/taikutsu/Taikutsu.kt
@@ -1,0 +1,9 @@
+package eu.kanade.tachiyomi.extension.es.taikutsu
+
+import eu.kanade.tachiyomi.multisrc.lectormoe.LectorMoe
+
+class Taikutsu : LectorMoe(
+    "Taikutsu",
+    "https://taikutsutl.lector.moe",
+    "es",
+)
